ci: Use cache / progress when local docker build
It's a bit annoying to try and work on these locally when the cache /
progress isn't being used so let's just set it so that those flags are
only valid when in CI directly.

`${CI}` is a default environment variable that's defined by actions
itself.

See https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/store-information-in-variables#default-environment-variables

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>

ghstack-source-id: e3e71103a7a0dfa20189c423f0c9e7f2f1fb9655
Pull Request resolved: https://github.com/pytorch/pytorch/pull/150418